scraper/requirements.txt
# Web Framework
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0

# HTTP Requests
requests==2.32.5
httpx==0.28.1
brotli==1.1.0

# HTML Parsing
beautifulsoup4==4.14.3
lxml==5.4.0
html5lib==1.1

# Selenium & Browser Automation
selenium==4.29.0
undetected-chromedriver==3.5.5
webdriver-manager==4.0.2

# Content Extraction & Cleaning
html2text==2020.1.16
readability-lxml==0.8.1
trafilatura==2.0.0; platform_system == "Darwin"
trafilatura==1.6.3; platform_system != "Darwin"
pdfminer.six==20221105

# Text Processing
python-multipart==0.0.6
charset-normalizer>=3.4.0,<4

# Utilities
python-dotenv==1.0.0
fake-useragent==1.4.0
tenacity==8.2.3
duckduckgo-search==4.1.1
setuptools>=75.0.0
